Add WorkspaceServer for integrated file operations and code execution
- Implemented file handling tools: read_file, write_file, list_directory, search_files, replace_lines_in_file, grep.
- Added code execution tools: execute_python, execute_typescript, execute_shell.
- Integrated SecureFileManager for path validation and file operations.
- Enhanced error handling and response formatting for better consistency.
- Introduced support for line range operations in file reading and replacing.
- Ensured sandboxed execution of code with timeout management.
